app/requirements.txt
fastapi 
uvicorn 
sqlalchemy 
pydantic
python-jose 
passlib
python-dotenv
python-multipart
psycopg2
